backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.0.2
openai==1.7.0
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1
